+++ bilby/gw/__init__.py
@@ -1,5 +1,8 @@
-from . import (calibration, conversion, detector, likelihood, prior, source,
-               utils, waveform_generator, result)
+from . import (calibration, conversion, cosmology, detector, likelihood, prior,
+               result, source, utils, waveform_generator)
+
 
 from .waveform_generator import WaveformGenerator
 from .likelihood import GravitationalWaveTransient
+
+