+++ bilby/gw/conversion.py
@@ -6,42 +6,48 @@ from pandas import DataFrame
 from ..core.utils import logger, solar_mass
 from ..core.prior import DeltaFunction, Interped
 from .utils import lalsim_SimInspiralTransformPrecessingNewInitialConditions
-
+from .cosmology import get_cosmology
 
 try:
-    from astropy.cosmology import z_at_value, Planck15
-    import astropy.units as u
+    from astropy import units
+    from astropy.cosmology import z_at_value
 except ImportError:
-    logger.warning("You do not have astropy installed currently. You will"
-                   " not be able to use some of the prebuilt functions.")
+    logger.debug("You do not have astropy installed currently. You will"
+                 " not be able to use some of the prebuilt functions.")
 
 
-def redshift_to_luminosity_distance(redshift):
-    return Planck15.luminosity_distance(redshift).value
+def redshift_to_luminosity_distance(redshift, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    return cosmology.luminosity_distance(redshift).value
 
 
-def redshift_to_comoving_distance(redshift):
-    return Planck15.comoving_distance(redshift).value
-
-
-@np.vectorize
-def luminosity_distance_to_redshift(distance):
-    return z_at_value(Planck15.luminosity_distance, distance * u.Mpc)
+def redshift_to_comoving_distance(redshift, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    return cosmology.comoving_distance(redshift).value
 
 
 @np.vectorize
-def comoving_distance_to_redshift(distance):
-    return z_at_value(Planck15.comoving_distance, distance * u.Mpc)
+def luminosity_distance_to_redshift(distance, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    return z_at_value(cosmology.luminosity_distance, distance * units.Mpc)
 
 
-def comoving_distance_to_luminosity_distance(distance):
-    redshift = comoving_distance_to_redshift(distance)
-    return redshift_to_luminosity_distance(redshift)
+@np.vectorize
+def comoving_distance_to_redshift(distance, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    return z_at_value(cosmology.comoving_distance, distance * units.Mpc)
 
 
-def luminosity_distance_to_comoving_distance(distance):
-    redshift = luminosity_distance_to_redshift(distance)
-    return redshift_to_comoving_distance(redshift)
+def comoving_distance_to_luminosity_distance(distance, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    redshift = comoving_distance_to_redshift(distance, cosmology)
+    return redshift_to_luminosity_distance(redshift, cosmology)
+
+
+def luminosity_distance_to_comoving_distance(distance, cosmology=None):
+    cosmology = get_cosmology(cosmology)
+    redshift = luminosity_distance_to_redshift(distance, cosmology)
+    return redshift_to_comoving_distance(redshift, cosmology)
 
 
 @np.vectorize
@@ -123,6 +129,23 @@ def convert_to_lal_binary_black_hole_parameters(parameters):
     converted_parameters = parameters.copy()
     original_keys = list(converted_parameters.keys())
 
+    if 'redshift' in converted_parameters.keys():
+        converted_parameters['luminosity_distance'] = \
+            redshift_to_luminosity_distance(parameters['redshift'])
+    elif 'comoving_distance' in converted_parameters.keys():
+        converted_parameters['luminosity_distance'] = \
+            comoving_distance_to_luminosity_distance(
+                parameters['comoving_distance'])
+
+    for key in original_keys:
+        if key[-7:] == '_source':
+            if 'redshift' not in converted_parameters.keys():
+                converted_parameters['redshift'] =\
+                    luminosity_distance_to_redshift(
+                        parameters['luminosity_distance'])
+            converted_parameters[key[:-7]] = converted_parameters[key] * (
+                1 + converted_parameters['redshift'])
+
     if 'chirp_mass' in converted_parameters.keys():
         if 'total_mass' in converted_parameters.keys():
             converted_parameters['symmetric_mass_ratio'] =\
@@ -198,14 +221,6 @@ def convert_to_lal_binary_black_hole_parameters(parameters):
             converted_parameters[angle] =\
                 np.arccos(converted_parameters[cos_angle])
 
-    if 'redshift' in converted_parameters.keys():
-        converted_parameters['luminosity_distance'] =\
-            redshift_to_luminosity_distance(parameters['redshift'])
-    elif 'comoving_distance' in converted_parameters.keys():
-        converted_parameters['luminosity_distance'] = \
-            comoving_distance_to_luminosity_distance(
-                parameters['comoving_distance'])
-
     added_keys = [key for key in converted_parameters.keys()
                   if key not in original_keys]
 
@@ -243,6 +258,18 @@ def convert_to_lal_binary_neutron_star_parameters(parameters):
     converted_parameters, added_keys =\
         convert_to_lal_binary_black_hole_parameters(converted_parameters)
 
+    for idx in ['1', '2']:
+        mag = 'a_{}'.format(idx)
+        if mag in original_keys:
+            tilt = 'tilt_{}'.format(idx)
+            if tilt in original_keys:
+                converted_parameters['chi_{}'.format(idx)] = (
+                    converted_parameters[mag] *
+                    np.cos(converted_parameters[tilt]))
+            else:
+                converted_parameters['chi_{}'.format(idx)] = (
+                    converted_parameters[mag])
+
     # catch if tidal parameters aren't present
     if not any([key in converted_parameters for key in
                 ['lambda_1', 'lambda_2', 'lambda_tilde', 'delta_lambda']]):
@@ -272,18 +299,6 @@ def convert_to_lal_binary_neutron_star_parameters(parameters):
             converted_parameters['lambda_1']\
             * converted_parameters['mass_1']**5\
             / converted_parameters['mass_2']**5
-
-    for idx in ['1', '2']:
-        mag = 'a_{}'.format(idx)
-        if mag in original_keys:
-            tilt = 'tilt_{}'.format(idx)
-            if tilt in original_keys:
-                converted_parameters['chi_{}'.format(idx)] = (
-                    converted_parameters[mag] *
-                    np.cos(converted_parameters[tilt]))
-            else:
-                converted_parameters['chi_{}'.format(idx)] = (
-                    converted_parameters[mag])
 
     added_keys = [key for key in converted_parameters.keys()
                   if key not in original_keys]